Updated variable type

+++ 0x01-variables_if_else_while/6-print_numberz.c
@@ -9,7 +9,7 @@
 
 int main(void)
 {
-	short a;
+	int a;
 
 	for (a = '0' ; a <= '9' ; a++)
 	{
